+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/screen/bestphoto/BestPhotoScreen.kt
@@ -1,0 +1,9 @@
+package band.effective.office.tv.screen.bestphoto
+
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun BestPhotoScreen() {
+    Text("Best Photo")
+}
