+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/screen/history/HistoryScreen.kt
@@ -1,0 +1,9 @@
+package band.effective.office.tv.screen.history
+
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun HistoryScreen() {
+    Text("History")
+}
