+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/screen/leaderIdEvets/LeaderIdEventScreen.kt
@@ -1,0 +1,9 @@
+package band.effective.office.tv.screen.leaderIdEvets
+
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun LeaderIdEventScreen() {
+    Text(text = "Events")
+}
