+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/screen/load/LoadCircle.kt
@@ -19,7 +19,7 @@ fun LoadCircle(modifier: Modifier, inactive: Int){
             val positions: List<Pair<Float, Float>> = getCords(Pair(size.width/2,size.height/2),size.width/2 )
             val colors: List<Color> = List(positions.size){if (inactive % positions.size == it) CharlestonGreen else Color.White}
             translate(left = -size.width/2) {
-                for (index in 0 until positions.size){
+                for (index in positions.indices){
                     translate(left = positions[index].first, top = positions[index].second) {
                         drawCircle(color = colors[index], radius = 5.dp.toPx())
                     }
