+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/ui/theme/Theme.kt
@@ -1,14 +1,12 @@
-package com.example.effecticetv.ui.theme
+package band.effective.office.tv.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.darkColors
 import androidx.compose.material.lightColors
 import androidx.compose.runtime.Composable
-import band.effective.office.tv.ui.theme.Purple200
-import band.effective.office.tv.ui.theme.Purple500
-import band.effective.office.tv.ui.theme.Purple700
-import band.effective.office.tv.ui.theme.Teal200
+import com.example.effecticetv.ui.theme.Shapes
+import com.example.effecticetv.ui.theme.Typography
 
 private val DarkColorPalette = darkColors(
     primary = Purple200,
